Include MolName when shuffling data in load_data()
The names of ligands weren't shuffled together with classes and data (if randomize=True), which could result in wrong names for ligands, for example, when printing ones with best scores. I changed the code a little so the names are shuffled in the same way as classes - inside the MolData list.